refactor: Integrate DI and remove mock repository implementations
This commit refactors the application to use dependency injection (DI) for providing repositories and removes the old mock data implementations.

- Deletes the mock repository implementations from `lib/data/repositories/` which are now obsolete.
- Updates `main.dart` to initialize the DI container (`initAppModule`) and provide repositories using `di.instance()` instead of direct instantiation.
- Removes the old `main.dart` from the `lib/ui/` directory.
- Adjusts Cubits (`HomeTopStoresCubit`, `HomeTrendingProductsCubit`, `UserCubit`, `CartCubit`, `HomeCategoriesCubit`) to correctly handle the `Result` wrapper by accessing the `.data` property.
- Simplifies the `CartRepository` by removing `clearCart` and `getCartItemCount` methods.
- Registers `OffersRepository` in the DI container.
- Corrects import paths in various UI files.
