+++ lib/di/injection_container.dart
@@ -1,4 +1,6 @@
 import 'package:get_it/get_it.dart';
+import 'package:sellio_mobile/data/repository_impl/offers_repository_impl.dart';
+import 'package:sellio_mobile/domain/repositories/offers_repository.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:internet_connection_checker/internet_connection_checker.dart';
 
@@ -168,6 +170,14 @@ void _initRepositoriesModule() {
         instance<CategoryLocalDataSource>(),
       ),
     );
+  }
+
+
+  if (!instance.isRegistered<OffersRepository>()) {
+    instance.registerLazySingleton<OffersRepository>(
+          () => OffersRepositoryImpl(),
+    );
+
   }
 
   if (!instance.isRegistered<ProductRepository>()) {
