+++ lib/presentation/screens/main/main_screen.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
 import '../../../core/design_system/widgets/bottom_nav_bar.dart';
+import '../account_screen.dart';
 import '../cart_screen.dart';
 import '../home/home_screen.dart';
 import '../customize_your_product_screen/CustomizeYourProductScreen.dart';
-import '../Account/account_screen.dart';
-import '../home/home_screen.dart';
 import '../thrift_screen.dart';
 
 class MainScreen extends StatefulWidget {
